requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
base58==2.1.1
bitsandbytes==0.42.0
blis==0.7.11
boto3==1.34.59
botocore==1.34.59
cachetools==5.3.3
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
coloredlogs==15.0.1
commonmark==0.9.1
confection==0.1.4
contourpy==1.1.1
cupy-cuda11x==12.3.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.8
datasets==2.19.1
diffusers==0.30.0
dill==0.3.8
diskcache==5.6.3
docker-pycreds==0.4.0
einops==0.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl#sha256=f72abb34bdf174876bd4267b29b2501677e605e0a251fdc56c163003182ed68b
exceptiongroup==1.2.0
fairscale==0.4.6
faiss-cpu==1.8.0
fastrlock==0.8.2
filelock==3.7.1
flash-attn==2.6.3
flatbuffers==24.3.25
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2023.12.2
fuzzywuzzy==0.18.0
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.17.1
google-auth==2.28.2
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.64.0
h5py==3.10.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.6
imageio==2.34.2
importlib_metadata==7.1.0
importlib_resources==6.1.3
iniconfig==2.0.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonnet==0.20.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.4
Levenshtein==0.25.1
llama_cpp_python==0.2.85
llvmlite==0.41.1
lmdb==1.4.1
Markdown==3.6
MarkupSafe==2.1.4
matplotlib==3.7.5
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
networkx==3.1
nltk==3.8.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.18.1
opencv-python-headless==4.10.0.84
packaging==23.2
pandas==2.0.3
pathlib_abc==0.1.1
pathtools==0.1.2
pathy==0.11.0
pillow==10.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
pluggy==1.4.0
pooch==1.8.2
preshed==3.0.9
promise==2.3
protobuf==3.20.3
psutil==5.9.8
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.4
pydantic_core==2.16.3
pygame==2.6.0
Pygments==2.17.2
PyMatting==1.1.12
pyparsing==3.1.2
pytest==8.1.1
python-dateutil==2.9.0.post0
python-Levenshtein==0.25.1
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
rake-nltk==1.0.6
rapidfuzz==3.9.5
referencing==0.35.1
regex==2023.12.25
rembg==2.0.58
requests==2.31.0
requests-oauthlib==2.0.0
rich==12.6.0
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.0
sacremoses==0.1.1
safetensors==0.4.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sentencepiece==0.1.99
sentry-sdk==1.42.0
setproctitle==1.3.3
shortuuid==1.0.13
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
spacy==3.7.4
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
statistic==0.1.6
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==1.1.0
textblob==0.17.1
thinc==8.2.3
threadpoolctl==3.2.0
tifffile==2023.7.10
tiktoken==0.5.2
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0
torch-tb-profiler==0.4.3
torch_geometric==2.5.3
torchvision==0.19.0
tqdm==4.66.1
traitlets==5.14.2
transformers==4.44.0
trimesh==4.4.4
triton==3.0.0
typer==0.4.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
wandb==0.17.0
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.3
xformers==0.0.27.post2
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
